Moved hook setup to scanbuilder
